[mpv] add high fps profile remove hwdec for yt
hwdec doesn't seem to reduce rendering time that much compared to
using less aggressive filters. So scalers are set to bilinear and deband
is disabled (by default would only be set for yt but maybe it could be
set manually and the profile would kick in if the framerate goes over 30
at some point in the video).
Only a proof that I run a potato