[nvim] simplify my nvim config
turned into a monster when it should not be so complex in $CURRENT_YEAR.
don’t even use a package manager and just rely on git submodules since
we already need them for these dotfiles anyway.
we’re missing none-ls but I’ll see if I have a need for it later.